% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMWrapper.R
\name{EMCluster}
\alias{EMCluster}
\title{Expectation Maximization Clustering}
\usage{
EMCluster(data = NULL, ...)
}
\arguments{
\item{data}{A Data set}

\item{\dots}{k: The number of Clusters}
}
\value{
A list of cluster labels and a returned object from init.EM
}
\description{
This function perform EM algorithm for model-based clustering of finite
mixture multivariate Gaussian distribution.The general purpose of clustering
is to detect clusters of data and to assign the data to the clusters.
}
\examples{
library(datasets)
data(iris)

rndSamples <- sample(nrow(iris),100)
trainData <- iris[rndSamples,]
testData <- iris[-rndSamples,]

clsut <- EMCluster(trainData[,1:4],3)
}
