% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMWrapper.R
\name{predict.EMCluster}
\alias{predict.EMCluster}
\title{EMCluster prediction function}
\usage{
\method{predict}{EMCluster}(object, ...)
}
\arguments{
\item{object}{A returned object of EMCluster}

\item{\dots}{New sample set}
}
\value{
A list of cluster labels
}
\description{
This function predicts the labels of the cluster for new data based on
cluster labels of the training set.
}
