% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalWrapper.R
\name{predict.hierarchicalCluster}
\alias{predict.hierarchicalCluster}
\title{hierarchicalCluster prediction function}
\usage{
\method{predict}{hierarchicalCluster}(object, ...)
}
\arguments{
\item{object}{A returned object of hierarchicalCluster function}

\item{\dots}{New samples set}
}
\value{
A list of cluster labels
}
\description{
This function predicts the labels of the cluster for new data based on
cluster labels of the training set.
}
