% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanWrapper.R
\name{predict.kmeansCluster}
\alias{predict.kmeansCluster}
\title{kmeansCluster prediction function}
\usage{
\method{predict}{kmeansCluster}(object, ...)
}
\arguments{
\item{object}{A returned object of kmeansCluster function}

\item{\dots}{New samples set}
}
\value{
A list of cluster labels
}
\description{
This function predicts the labels of the cluster for new data based on
cluster labels of the training set.
}
