% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamWrapper.R
\name{predict.pamCluster}
\alias{predict.pamCluster}
\title{pamCluster prediction function}
\usage{
\method{predict}{pamCluster}(object, ...)
}
\arguments{
\item{object}{A returned object of pamCluster function}

\item{\dots}{New samples set}
}
\value{
A list of cluster labels
}
\description{
This function predicts the labels of the cluster for new data based on
cluster labels of the training set.
}
